Backchannel package refactored
