+++ components/org.wso2.carbon.identity.oidc.session/src/main/java/org/wso2/carbon/identity/oidc/session/logout/backchannel/ClaimProviderImpl.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.identity.oidc.session.logout.backchannellogout;
+package org.wso2.carbon.identity.oidc.session.logout.backchannel;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
