+++ components/org.wso2.carbon.identity.oidc.session/src/main/java/org/wso2/carbon/identity/oidc/session/logout/backchannel/LogoutRequestSender.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.carbon.identity.oidc.session.logout.backchannellogout;
+package org.wso2.carbon.identity.oidc.session.logout.backchannel;
 
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.logging.Log;
